ref: move `bbye` to `bufferline.bbye`
