+++ lua/bufferline/state.lua
@@ -9,6 +9,7 @@ local utils = require'bufferline.utils'
 local Buffer = require'bufferline.buffer'
 local Layout = require'bufferline.layout'
 local animate = require'bufferline.animate'
+local bbye = require'bufferline.bbye'
 local len = utils.len
 local is_nil = utils.is_nil
 local index_of = utils.index_of
@@ -576,7 +577,7 @@ local function close_all_but_current()
   local buffers = m.buffers
   for _, number in ipairs(buffers) do
     if number ~= current then
-      require'bbye'.delete('bdelete', false, bufname(number))
+      bbye.delete('bdelete', false, bufname(number), nil)
     end
   end
   m.update()
@@ -586,7 +587,7 @@ local function close_all_but_pinned()
   local buffers = m.buffers
   for _, number in ipairs(buffers) do
     if not is_pinned(number) then
-      require'bbye'.delete('bdelete', false, bufname(number))
+      bbye.delete('bdelete', false, bufname(number), nil)
     end
   end
   m.update()
@@ -597,7 +598,7 @@ local function close_all_but_current_or_pinned()
   local current = nvim.get_current_buf()
   for _, number in ipairs(buffers) do
     if not is_pinned(number) and number ~= current then
-      require'bbye'.delete('bdelete', false, bufname(number))
+      bbye.delete('bdelete', false, bufname(number), nil)
     end
   end
   m.update()
@@ -609,7 +610,7 @@ local function close_buffers_left()
     return
   end
   for i = idx, 1, -1 do
-    require'bbye'.delete('bdelete', false, bufname(m.buffers[i]))
+    bbye.delete('bdelete', false, bufname(m.buffers[i]), nil)
   end
   m.update()
 end
@@ -620,7 +621,7 @@ local function close_buffers_right()
     return
   end
   for i = len(m.buffers), idx, -1 do
-    require'bbye'.delete('bdelete', false, bufname(m.buffers[i]))
+    bbye.delete('bdelete', false, bufname(m.buffers[i]), nil)
   end
   m.update()
 end
